Add evaluation scripts for submission
