--- scripts/g3_lerobotpi/challenge_widowx.sh
@@ -1,15 +1,0 @@
-#!/bin/bash
-
-
-ckpt=$1
-action_ensemble=${2:-0}
-
-echo "Evaluating checkpoint: $ckpt"
-echo "Action ensemble: $action_ensemble"
-
-args=( --ckpt-path "$ckpt" )
-[[ "${action_ensemble,,}" =~ ^(true)$ ]] && args+=( --action-ensemble )
-
-args+=( --save-path-suffix "$(date +%Y%m%d_%H%M%S)" )
-
-python scripts/g3_lerobotpi/challenge_widowx.py "${args[@]}"
