+++ scripts/group3/evaluate_bridge.py
@@ -3,26 +3,28 @@ import time
 
 import numpy as np
 
-from scripts.g3_lerobotpi.g3_configuration_pi0 import G3PI0Config
-from scripts.g3_lerobotpi.g3_pi0_or_fast import G3LerobotPiFastInference
+from scripts.group3.g3_configuration_pi0 import G3PI0Config
+from scripts.group3.g3_pi0_or_fast import G3LerobotPiFastInference
 from simpler_env.evaluation.bridge_tasks import (
     widowx_task1_pick_object,
     widowx_task2_stack_cube,
     widowx_task3_put_object_on_top,
     widowx_task4_put_object_in_basket
 )
-from simpler_env.evaluation.evaluate import calculate_robust_score
 
 def parse_args():
     parser = argparse.ArgumentParser(description="Run Comprehensive ManiSkill2 Evaluation")
     parser.add_argument("--ckpt-path", type=str, required=True, help="Path to the checkpoint to evaluate.")
     parser.add_argument("--control-freq", type=int, default=5, help="Set control frequency (default->5)")
-    parser.add_argument("--action-ensemble", action="store_true", help="Use action ensemble if set.")
-    parser.add_argument("--save-path-suffix", type=str, default="", help="Suffix to add to the save path.")
     return parser.parse_args()
 
 
 if __name__ == "__main__":
+    N_ACTION_STEPS = 1
+    ACTION_ENSEMBLE_TEMP = 0.8
+    ACTION_ENSEMBLE = False
+    STICKY_ACTION = False
+
     args = parse_args()
     ckpt_path = args.ckpt_path
 
@@ -30,7 +32,10 @@ if __name__ == "__main__":
         saved_model_path=ckpt_path,
         policy_setup="widowx_bridge",
         action_scale=1.0,
-        action_ensemble=args.action_ensemble,
+        action_ensemble_temp=ACTION_ENSEMBLE_TEMP,
+        action_ensemble=ACTION_ENSEMBLE,
+        sticky_action=STICKY_ACTION,
+        n_action_steps=N_ACTION_STEPS,
     )
 
     print("Policy initialized. Starting evaluation...")
@@ -44,19 +49,10 @@ if __name__ == "__main__":
 
     final_scores = []
     for task in tasks:
-        s = time.time()
-
-        # Evaluate a task
         cur_scores = task(
-            env_policy=policy, ckpt_path=args.ckpt_path + f"_{args.save_path_suffix}", control_freq=args.control_freq
+            env_policy=policy, ckpt_path=args.ckpt_path, control_freq=args.control_freq
         )
         final_scores += cur_scores
 
-        # Log the results
-        print(f"Task: {task.__name__}")
-        print(f"Time: {(time.time() - s)/60:.2f} min")
-        print(f"Success Rate: {np.mean(cur_scores)*100:.2f}% ({np.sum(cur_scores)}/{np.prod(np.shape(cur_scores))})")
-
     print("\nEvaluation finished.")
-    print(f"Final Success Rate: {np.mean(final_scores)*100:.2f}%")
-    print(f"Final calculated scores: {calculate_robust_score(final_scores)}")
+    print(f"Final calculated scores: {final_scores}")
